backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
websockets==12.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
apscheduler==3.10.4
numpy==1.26.4
scipy==1.13.0
python-dotenv==1.0.1
aiofiles==23.2.1
pyyaml==6.0.1
pytz==2024.1
astral==3.2
python-multipart==0.0.9
